Create grabs_tags.py
chapter 2 page 15 
context: the url page contains a story. The names of the characters are in green. Where as the lines spoken are in red. 

First we grab the content of the url and create a BeautifulSoup object
Second, we use findAll method of bs4 to find all of the span tags which has the class green 
Last, we loop through the list using the get_text() method